requirements.txt
alembic==1.13.2
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
langchain-chroma==0.2.2
chromadb==0.6.3
yandex-chain==0.0.9
pydantic==2.11.7
pydantic-settings==2.1.0
joblib==1.5.2
loguru==0.7.3
langchain-community==0.2.4
langchain-core==0.2.43
gunicorn==21.2.0
atlassian-python-api
beautifulsoup4
pandas
numpy